Assignment1
 In String Methods:
With rstrip(char), a character is removed from the end of a string. 
strip(char): this function eliminates the given character from the string's two ends. 
isalnum(): Verifies that every character is an alphanumeric character. 
isspace(): Verifies if any character is a whitespace character. 
split(): divides a string according to whitespace into a list.

Regular Expressions:
importing regular expression as regex
where it is used in searching, processing and extracting the unstructured data.
The offered code's tokenize function takes the input string and extracts using a regular expression pattern and the re.findall() method. After that, a list containing the extracted is printed.