Add files via upload
This code is a comprehensive data analysis and visualization project using Python libraries such as pandas, seaborn, nltk, and wordcloud. The data is sourced from a CSV file containing UN General Debates.

The initial part of the code imports necessary libraries and reads the CSV file into a pandas DataFrame (df). The DataFrame's structure and content are explored using various methods like head(), shape, columns, dtypes, info(), and describe() to get an overview of the dataset.
A text processing pipeline is established using the tokenize and remove_stop functions to preprocess the textual data. The pipeline is then applied to the 'Book Name' column of the DataFrame, resulting in a new 'tokens' column that contains the preprocessed tokens.
The frequency of the tokens is calculated using the Counter class from the collections module. This frequency count is stored in the counter object and subsequently converted into a DataFrame (freq_df). The most common tokens are then visualized using bar plots to display the top 15 or 10 words by frequency.
Finally, the code uses the WordCloud library to create visual representations of the word frequencies. One word cloud focuses on the text from the year 2015 and the country USA, while the other represents the frequency of words across the entire dataset.