Add files via upload
The required libraries, such as NLTK for natural language processing, regex for text processing, and Pandas for data manipulation, are imported at the start of the code. Next, a Pandas DataFrame is filled with a dataset including UN general discussions. After that, a set of the English stopwords is downloaded and saved. A pipeline for text pre-processing is built that removes stopwords, tokenizes the text with regex, and transforms it to lowercase. The 'text' column of the DataFrame is subjected to this pipeline, and the output tokens are kept in a new column called 'tokens'.
Second, the code uses the Counter class from the collections module to determine how often each token appears in the 'tokens' column. The tokens and their frequencies are assembled into a DataFrame, and tokens with frequencies smaller than one are removed. Next, using Matplotlib and the WordCloud module, a word cloud of the most frequent tokens is produced.
The function then determines each token's inverse document frequency (IDF). A custom function that counts the number of documents in which each token appears is used to compute the IDF. To calculate the TF-IDF, the IDF values are then multiplied by each token's term frequency (TF). We create another word cloud with the TF-IDF values.
And last, the code concentrates on particular terms associated with international affairs, such 'terrorism,' 'nuclear,' war,'oil,'syria,'afghanistan,'refugees,'israel,'peacekeeping,'humanitarian,'climate,'soviet,'russia,'sustainable,' and 'egypt'. These terms appear in the "tokens" column once, and their counts are totaled and sorted by year. Using Seaborn, a heatmap is then made to show how frequently certain keywords have occurred over time. To improve the heatmap's contrast, the frequency data are square-root transformed and adjusted based on the total amount of tokens issued annually.