Add files via upload
Importing a dataset from a CSV file with Billboard song data is the first step in the study. The text data in the 'Song Name' column is pre-processed by tokenizing, eliminating stopwords, and changing the text's case once it has been imported. Following the counting of the most frequent tokens in the 'Song Name' column, a word cloud is created to represent these frequencies.
The next step is to assess the significance of each word in the dataset using the Term Frequency-Inverse Document Frequency (TF-IDF) method. To determine how rare each token is throughout the documents, the IDF values for the tokens are calculated. Each token's TF-IDF score is obtained by multiplying the TF and IDF values; a new word cloud is then created using this information.
Then, utilizing the 'Last Week' column from the dataset, a second function is constructed to group the counts by year. The first method counts certain terms within the token list. Seaborn is used to create a heatmap that shows the relative frequency of particular terms over time.
The following keywords were selected to be included in the heatmap: "Like That," "Texas Hold," "Beautiful Things," "Lose Control," "Beyonce," "Benson Boone," "Tyler Hubbard," "Teddy Swims," "Everyday Hustle," "home," "Good Neighbors," "Tyler Hubbard," "Jordan Davis," and "Hozier." The relative frequency of these terms is depicted annually in a heatmap, where the color intensity corresponds to the keyword's frequency.