Add files via upload
The names of the most popular entries from the Spotify subreddit are scraped by this script using the PRAW (Python Reddit API Wrapper) package. The script retrieves the titles of the most popular entries from the "spotify" subreddit after configuring the PRAW Reddit instance using the given client credentials and a custom user agent. After retrieving unique titles, it uses the TextBlob library to clean up the content by deleting RT, hashtags, emojis, and hyperlinks. Finally, it determines the subjectivity and polarity of each title. The sentiment of the titles is categorized as "Negative," "Neutral," or "Positive" using the subjectivity and polarity values, and then inserted as a new column in the DataFrame. Ultimately, it creates a word cloud from the sentiment scores and uses a bar plot to represent them.the titles that have been cleaned. The sentiment score distribution is displayed in a bar plot, and the most common terms in the titles are visually represented in a word cloud.