Add files via upload
This code performing sentiment analysis using the Bing Liu lexicon on an Arts, Crafts, and Sewing product review dataset. It imports necessary libraries, reads the dataset, and observes a sample of 10 rows. Then, it retrieves the Bing Liu lexicon from NLTK, creates a dictionary of positive and negative words, and defines a function to compute sentiment scores based on this lexicon. After replacing null values in the 'text' column with 'no review', it calculates sentiment scores for each review text. Finally, it displays the first 10 rows of the 'overall', 'text', and 'Bing_Liu_Score' columns and computes the mean sentiment score grouped by the 'overall' rating.