Add files via upload
Using pandas for data manipulation, nltk for text processing, and regex for tokenization, the presented code first imports the required libraries. The text is preprocessed by tokenizing, deleting common stopwords, and converting the text to lowercase after it has read a dataset of UN General Debate speeches. Following a count of the tokens, a DataFrame containing their frequencies is created. A word cloud is created to display the most frequently occurring terms using the WordCloud and matplotlib tools. To get TF-IDF scores, which are utilized to produce a more illuminating word cloud, the code moreover computes the Inverse Document Frequency (IDF) for each token and multiplies it by the token frequencies. Furthermore defined functions are ones that count certain keywords found in the text and organize the counts by year. Lastly,  that shows the relative frequency of keywords over time is made using seaborn, giving insights into the phrases' temporal occurrence in UN talks. To extract and illustrate important themes from the dataset, this all-inclusive method combines text preparation, frequency analysis, TF-IDF computation, and data visualization.