Add files via upload
This code with with includes python libraries like pandas, seaborn, nltk, and wordcloud and to provide full data analysis the source of the data is CSV file.the first it reads the CSV file to provide summary of the dataset using head(), shape, columns, dtypes, info(), and describe(). The next steps involve cleaning and preparing the data, filling in any missing values in the "speaker" and "position" columns with "unknown," and changing all text-based columns to uppercase to maintain consistency. A text processing pipeline is created with the tokenize and remove_stop functions. The tokens are then contained in a new "tokens" column created by applying the pipeline to the DataFrame "text" column. Using the Counter class from the collections module the frequency of the tokens is determined. . After being saved in the counter object, this frequency count is transformed into a DataFrame (freq_df). The top 15 or 10 terms by frequency are then displayed using bar plots. The word frequencies are shown by the code by utilizing the WordCloud library. While the other word cloud shows the frequency of words throughout the entire dataset, the first word cloud concentrates on the material from 2015 and the United States.