Add files via upload
The provided code uses the PRAW library to communicate with Reddit's API while conducting a thorough analysis of the trending topics from the "bitcoin" subreddit. To use the Reddit API, the Reddit instance is first configured with a custom user agent and client credentials. From the "bitcoin" subreddit, the code collects the unique titles of every hot entry, which it then keeps in a DataFrame. Next, using regular expressions and TextBlob, the text is cleaned to get rid of mentions, hashtags, RT tags, hyperlinks, and emoticons. Sentiment analysis is then carried out using TextBlob to ascertain the subjectivity and polarity of each title following text cleaning. Three sentiment categories "Positive," "Neutral," and "Negative"—are created from the polarity scores. Seaborn and Matplotlib are used in data visualization to plot the Distribution of emotion score and provide a word cloud visualization of the article titles pertaining to bitcoin. The ensuing visuals provide information on the mood and popular subjects within the "bitcoin" subreddit.